+++ spec/spec_helper.rb
@@ -2,6 +2,7 @@ $LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
 require "pry-byebug"
 require "api_client_base"
 require "virtus-matchers"
+require "dry-validation"
 require 'vcr'
 
 VCR.configure do |c|